Add static mesh colliders
Static mesh colliders are immovable i.e. once instantiated, they cannot
be moved either manually or by any other physics object colliding with
it. See Bullet manual for more information.

Currently we are only allowing the user to supply .obj files as mesh
files for the mesh collider because we want to encourage simple models
to be used as mesh colliders. They do not need animations, materials,
textures, skeletons etc. and .obj is apt for that. We do not use the
.mtl file data which inevitably gets loaded by Assimp. All that extra data is
loaded and put back as is.