Add collision masks and collision groups
These work exactly the same as described in the bullet 3D user manual. A
collision group is the property of which groups a collider belongs to. A
collision mask is the property of which groups a collider is interested
in colliding into.

Collision only happens if the group of each colliding object in a
collision pair contains the collision mask of the other colliding object
in the collision pair.