Fix dialogue graph code to follow order of appearance
Implements this by maintaining a stack of the dialogue nodes which can
be returned to in case the answer nodes of QuestionNode reaches the
terminal. QuestionNode pushes its next chronological node into the stack
before it order the dialogue to move on to the answer nodes. If the
answer nodes hit a stop, the stack is checked back for any nodes that
can be returned to. If no such return values are available, the dialogue
stops.