Remove ambiguity for NVIDIA FXAA license + cut link to Wine's LGPL code
Add in NVIDIA's FXAA implementation licensing details in THIRDPARTY.md
(was probably left during last refactor of all third party deps)

Also remove an unnecessary link to Wine's LGPL source code from a comment.
We are removing it just because Wine is LGPL and leaving the link in could would get
undeserving attention because the amount of implementation detail
matching in our code base with the code in the link to Wine's codebase
was just general a utility function, which we have now replaced with a
completely original implementation which can only work for our use case
and not in any general case which Wine might be requiring.

Hence, we end any sort of references to all LGPL'd Wine source code
from any code belonging to SDSLabs in this repository.